galeri modülü
galeri modülü bitti

+++ panel/application/controllers/Galleries.php
@@ -19,8 +19,7 @@ class Galleries extends CI_Controller {
 			redirect(base_url("login"));
 		}
 	}
-	public function index()
-	{
+	public function index(){
 		$viewData = new stdClass();
 		// veri tabanından verilerin getirilmesi
 		$items = $this->gallery_model->get_all(
@@ -34,7 +33,6 @@ class Galleries extends CI_Controller {
 
 		$this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index", $viewData);
 	}
-
 
 	public function new_form(){
 
@@ -59,7 +57,97 @@ class Galleries extends CI_Controller {
 		$this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index", $viewData);
 	}
 
+	public function update_gallery_video_form($id){
 
+		$viewData = new stdClass();
+
+		/** Tablodan Verilerin Getirilmesi.. */
+		$item = $this->video_model->get(
+			array(
+				"id"    => $id,
+			)
+		);
+
+		/** View'e gönderilecek Değişkenlerin Set Edilmesi.. */
+		$viewData->viewFolder = $this->viewFolder;
+		$viewData->subViewFolder = "video/update";
+		$viewData->item = $item;
+
+		$this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index", $viewData);
+	}
+
+    public function gallery_video_update($id, $gallery_id){
+
+        $this->load->library("form_validation");
+
+        // Kurallar yazilir..
+        $this->form_validation->set_rules("url", "Video URL", "required|trim");
+        $this->form_validation->set_rules("title", "Video Adı", "required|trim");
+
+        $this->form_validation->set_message(
+            array(
+                "required"  => "<b>{field}</b> alanı doldurulmalıdır"
+            )
+        );
+
+        $validate = $this->form_validation->run();
+
+        if($validate){
+
+            $update = $this->video_model->update(
+                array(
+                    "id"    => $id
+                ),
+                array(
+                    "url"   => $this->input->post("url"),
+                    "title"   => $this->input->post("title"),
+                )
+            );
+
+            // TODO Alert sistemi eklenecek...
+            if($update){
+
+                $alert = array(
+                    "title" => "İşlem Başarılı",
+                    "text" => "Kayıt başarılı bir şekilde güncellendi",
+                    "type"  => "success"
+                );
+
+            } else {
+
+                $alert = array(
+                    "title" => "İşlem Başarısız",
+                    "text" => "Güncelleme sırasında bir problem oluştu",
+                    "type"  => "error"
+                );
+
+            }
+
+            $this->session->set_flashdata("alert", $alert);
+
+            redirect(base_url("galleries/upload_form/$gallery_id"));
+
+        } else {
+
+            $viewData = new stdClass();
+
+            /** Tablodan Verilerin Getirilmesi.. */
+            $item = $this->video_model->get(
+                array(
+                    "id"    => $id,
+                )
+            );
+
+            /** View'e gönderilecek Değişkenlerin Set Edilmesi.. */
+            $viewData->viewFolder = $this->viewFolder;
+            $viewData->subViewFolder = "video/update";
+            $viewData->form_error = true;
+            $viewData->item = $item;
+
+            $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index", $viewData);
+        }
+
+    }
 	public function save(){
 
 		$this->load->library("form_validation");
@@ -448,10 +536,14 @@ class Galleries extends CI_Controller {
 				$modelName = "image_model";
 				break;
 
-
 				case 'file':
 				$modelName = "file_model";
 				break;
+
+				case 'video':
+				$modelName = "video_model";
+				break;
+
 				default:
 				# code...
 				break;
@@ -479,12 +571,16 @@ class Galleries extends CI_Controller {
 			$modelName = "image_model";
 			break;
 
-
 			case 'file':
 			$modelName = "file_model";
 			break;
+	
+			case 'video':
+			$modelName = "video_model";
+			break;
+
 			default:
-		# code...
+			# code...
 			break;
 		}
 
@@ -663,8 +759,12 @@ class Galleries extends CI_Controller {
 			
 			case 'file':
 			$modelName = "file_model";
-
 			break;
+
+			case 'video':
+			$modelName = "video_model";
+			break;
+
 			default:
 				# code...
 			break;
@@ -680,7 +780,7 @@ class Galleries extends CI_Controller {
 
 		$delete = $this->$modelName->delete(
 			array(
-				"id" => $id,
+				"id" => $id
 			)
 		);
 
@@ -688,7 +788,7 @@ class Galleries extends CI_Controller {
 
 		if ($delete) {
 
-			unlink("$fileName->url");
+		if($gallery_type != "video"){ unlink("$fileName->url"); }
 
 			redirect(base_url("galleries/upload_form/$gallery_id"));
